feat: integrate react-router-dom for routing and introduce stats page
- Added react-router-dom to manage application routing.
- Refactored App component to include Router, NavBar, and Routes for Home and Stats pages.
