+++ package.json
@@ -22,6 +22,7 @@
     "posthog-js": "^1.234.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.6.0",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
